committers now only perform signature header validation and creator signature verification for config envelopes.
This was done because bftsmart ordering service creates configuration envelopes at the ordering nodes, and these cannot sign the
envelopes because otherwise they would produce different envelopes with different signatures that would not
match up at the frontends. But since they arrive within blocks that are signed by a Byzantine majority of nodes
(which are the ones that produce the config envelopes), it is safe to not perform header and signature verification
for config envelopes.